used correct script name
