+++ addObjectsToScene.mjs
@@ -5,7 +5,7 @@ import { loadBreak } from './model_inits/break_init.mjs';
 import { loadSeat } from './model_inits/car_seat_init.mjs';
 import { loadDisplay } from './model_inits/display_init.mjs';
 import { loadGasPedal } from './model_inits/gas_pedal_init.mjs';
-import { loadClutchPedal } from './model_inits/clutch_pedal.mjs';
+import { loadClutchPedal } from './model_inits/clutch_pedal_init.mjs';
 import { addObject } from './utils';
 
 export const addObjectsToScene = async (scene, gui) => {
